+++ embedded-io/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+- Increased MSRV to 1.83 due to `core::error::Error` & mutable reference in const context.
+
 ## 0.7.1 - 2025-09-30
 
 - Do not require `Read` and `Write` to be implemented for `ReadReady` and `WriteReady`.